Add seastar_kafka_client to scylla submodules
This commit adds the seastar_kafka_client lib to the list of Scylla
submodules. As for right now it points to the master branch
at github.com/haaawk/seastar-kafka-client.git and it should
stay that way until the lib is moved or forked into the ScyllaDB github.

To work correctly with the lib, we add a new include directory
in configure.py